Move IngressFromManifest/IngressToManifest to ingress e2e fw
