+++ test/e2e/framework/ingress/ingress_utils_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2020 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package manifest
+package ingress
 
 import (
 	"io/ioutil"
@@ -36,11 +36,11 @@ func TestIngressToManifest(t *testing.T) {
 	ingPath := filepath.Join(tmpDir, "ing.yaml")
 
 	// Write the ingress to a file and ensure that there is no error.
-	if err := IngressToManifest(ing, ingPath); err != nil {
+	if err := ingressToManifest(ing, ingPath); err != nil {
 		t.Fatalf("Error in creating file: %s", err)
 	}
 	// Writing it again should not return an error.
-	if err := IngressToManifest(ing, ingPath); err != nil {
+	if err := ingressToManifest(ing, ingPath); err != nil {
 		t.Fatalf("Error in creating file: %s", err)
 	}
 }
